requirements.txt
openai
pygithub
tweepy
python-telegram-bot
langchain
langgraph
python-dotenv
logging
telebot
tavily-python
langchain-deepseek